+++ CHANGES.md
@@ -2,9 +2,9 @@
 
 ## [NEXT]
 
-CHANGES
+### Changed
 
-* Breaking: removed actix feature `http`. Actix support for http was moved solely to actix-http and actix-web crates.
+* Feature `http` was removed. Actix support for http was moved solely to actix-http and actix-web crates.
 
 ## [0.9.1] 2020-??-??