+++ CHANGES.md
@@ -10,6 +10,8 @@
 
 * Allow return of any `T: 'static` on `ResponseFuture`. (#343)
 
+* Allow return of any `T: 'static` on `ResponseActFuture`. (#310)
+
 ## [0.9.0] 2019-12-20
 
 ### Fixed